requirements.txt
torch
torchvision
torchaudio
opencv-python-headless
imageio
imageio-ffmpeg
einops
safetensors
accelerate
pyyaml
